Extract peer-seed color calculation out
